requirements.txt
# AutoPost - Dependências
# ============================================================

# Instagram API não-oficial
instagrapi>=2.0.0

# Processamento de imagens
Pillow>=10.0.0

# Google Drive API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Agendamento
APScheduler>=3.10.0
pytz>=2023.3

# Configuração
python-dotenv>=1.0.0

# Validação
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilidades
httpx>=0.25.0
aiofiles>=23.2.0
